pyproject.toml
[tool.poetry]
name = "tragicentireapplicationprogrammer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
youtube-dl = "^2020.7.28"
async-timeout = "^3.0.1"
PyNaCl = "^1.4.0"
flask = "^1.1.2"
qrcode = "^6.1"
selenium = "^3.141.0"
python-decouple = "^3.4"
pafy = "^0.5.5"
pypng = "^0.0.21"
pytube = "^11.0.2"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
